docs: fix mkdocs `.page` issue
This commit is to fix all the issue w.r.t. mkdocs.

`.page` is excluded so it will not be applied in the top tabs.

The old solution to maintain tab order is to add nav directly
in mkdocs.yaml, which will disble awesome-pages. same-dir also
is not compatible with awesome-pages.

Meanwhile, the non-existing pages mention in
quick-start/start-quide/.pages cause awesome-pages not to work as
expected. So I comment them out.

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>